LibTLS: Rework method names and arrangement in cpp files
This commit only moves and renames methods. The code hasn't changed.
